feat(FN-1855): refactor file structure
