+++ libs/common/src/index.ts
@@ -1,5 +1,5 @@
 export * from './constants';
 export * from './helpers';
 // './sql-db-connection' should not be exported from here - see sql-db.md for details
-export * from './sql-db-entity';
+export * from './sql-db-entities';
 export * from './types';
