+++ libs/common/src/sql-db-entities/base-entities/index.ts
@@ -1,0 +1,1 @@
+export { AuditableEntity } from './auditable-entity';
