+++ libs/common/src/sql-db-entities/partial-entities/index.ts
@@ -1,3 +1,2 @@
-export { AuditableEntity } from './auditable-entity';
 export { MonthAndYearPartial } from './month-and-year';
 export { ReportPeriodPartial } from './report-period';
