+++ libs/common/src/sql-db-entities/partial-entities/report-period.ts
