+++ libs/common/src/sql-db-entities/utilisation-data.ts
@@ -1,7 +1,7 @@
 import { Column, Entity, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
 import { UtilisationReportEntity } from './utilisation-report';
 import { Currency } from '../types';
-import { AuditableEntity } from './partials';
+import { AuditableEntity } from './base-entities';
 
 // TODO FN-2183 - should this name maybe refer to `payments`, `feeRecords`, or something else?
 @Entity('UtilisationData')
